Remove custom code imports and extension writing from Java project
