+++ Templates/Java/requests_extensions/EntityRequestBuilder.java.tt
@@ -17,5 +17,5 @@
      */
     public <#=TypeRequestBuilder(c)#>(final String requestUrl, final <#=IBaseClientType()#> client, final java.util.List<Option> requestOptions) {
         super(requestUrl, client, requestOptions);
-    }<#=AddCustomCode(TypeRequestBuilder(c))#>
+    }
 }
