+++ Templates/Java/requests_extensions/IEntityRequestBuilder.java.tt
@@ -7,5 +7,4 @@
 <#=CreatePackageDef(host)#>
 
 <#=CreateInterfaceDef(ITypeRequestBuilder(c), IBaseTypeRequestBuilder(c))#>
-<#=AddCustomCode(ITypeRequestBuilder(c))#>
 }
