+++ src/GraphODataTemplateWriter/CodeHelpers/Android/CustomImports.cs
@@ -1,4 +1,8 @@
-﻿// Copyright (c) Microsoft Corporation.  All Rights Reserved.  Licensed under the MIT License.  See License in the project root for license information.
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
 
 namespace Microsoft.Graph.ODataTemplateWriter.CodeHelpers.Android
 {
@@ -8,8 +12,8 @@ namespace Microsoft.Graph.ODataTemplateWriter.CodeHelpers.Android
      * The field name should be equal to the name of the class you wish to edit
      * This will be used in BaseModel.template.tt in the AddCustomImports() method
      */
-    public class CustomImports : Java.CustomImports
+    public class CustomImports
     {
-        
+        public const string UploadSession = "import com.microsoft.graph.concurrency.ChunkedUploadProvider;\r\n";
     }
 }
