+++ src/GraphODataTemplateWriter/GraphODataTemplateWriter.csproj
@@ -59,13 +59,11 @@
     <Compile Include="CodeHelpers\Android\CustomImports.cs" />
     <Compile Include="CodeHelpers\Android\CustomMethods.cs" />
     <Compile Include="CodeHelpers\Android\CustomOverwrites.cs" />
-    <Compile Include="CodeHelpers\Java\CustomImports.cs" />
-    <Compile Include="CodeHelpers\Java\CustomMethods.cs" />
-    <Compile Include="CodeHelpers\Java\CustomOverwrites.cs" />
     <Compile Include="CodeHelpers\CodeWriterBase.cs" />
     <Compile Include="CodeHelpers\CSharp\CodeWriterCSharp.cs" />
     <Compile Include="CodeHelpers\CSharp\TypeHelperCSharp.cs" />
     <Compile Include="CodeHelpers\Java\CodeWriterJava.cs" />
+    <Compile Include="CodeHelpers\Java\CustomOverwrites.cs" />
     <Compile Include="CodeHelpers\Java\TypeHelperJava.cs" />
     <Compile Include="CodeHelpers\TypeScript\CodeWriterTypeScript.cs" />
     <Compile Include="CodeHelpers\TypeScript\TypeHelperTypeScript.cs" />
